# services/linkforge-gateway/src/main/resources/static/openapi/linkforge-public-api-v1.yaml
openapi: 3.1.0
info:
  title: LinkForge Public API
  version: 1.0.0
  description: Gateway 对外公开的用户、域名、短链、分组及跳转接口。
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
  /api/v1/users:
    post:
      tags: [Users]
      operationId: registerUser
      summary: 注册用户
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [Users]
      operationId: listUsers
      summary: 分页查询用户
      parameters:
        - $ref: '#/components/parameters/UserStatus'
        - $ref: '#/components/parameters/UserSearch'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 用户页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    TraceId:
      description: 服务端生成的请求关联标识
      schema:
        type: string
        maxLength: 64
  parameters:
...
    UserStatus:
      name: status
      in: query
      schema:
        type: string
        enum: [DEACTIVATED, ACTIVE, SECURITY_SUSPENDED]
    UserSearch:
      name: q
      in: query
      schema:
        type: string
        maxLength: 64
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: 请求参数不合法
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: 缺少或无法验证访问令牌
      headers:
        WWW-Authenticate:
          schema:
            type: string
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: 当前主体无权执行操作
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Conflict:
      description: 当前状态或唯一约束冲突
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, detail]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        traceId:
          type: string
        reasonCode:
          type: string
        decisionId:
          type: string
          format: uuid
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          pattern: '^[a-zA-Z0-9_]{4,64}$'
        password:
          type: string
          minLength: 6
          maxLength: 64
          writeOnly: true
        email:
          type: [string, 'null']
          format: email
          maxLength: 128
    RegisterResponse:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: [string, 'null']
          format: email
    User:
      type: object
      required: [id, username, status, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        username: {type: string}
        email: {type: [string, 'null'], format: email}
        realName: {type: [string, 'null']}
        status: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    UserPage:
      $ref: '#/components/schemas/CursorPageUsers'
    CursorPageUsers:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        nextCursor:
          type: [string, 'null']
        hasMore:
          type: boolean
...